--- src/config/config.ts
@@ -1,3 +1,0 @@
-/* istanbul ignore file */
-export * from './base-config'
-export * from './common'
