+++ src/index.ts
@@ -3,7 +3,6 @@ export * from './fs/fs'
 export * from './utils/utils'
 export * from './cli/cli'
 export * from './plugins/plugins'
-export * from './config/config'
 export * from './common/common'
 export * from './esbuild/esbuild'
 export * from './libs'
